#119 - Polishing.
Changed the update clause to set the the entire principal the reflect a domain object design that's closer to what might be set up if Spring Data auditing is used, too.

Made use of Lombok in the domain objects where possible. Reduced visibility of repository interfaces to package scope. Minor test case cleanups.