#197 - Upgraded to Spring Boot 1.4 RC1.
Tweaked the output folder for Querydsl type generation to avoid running into a bug in the Maven compiler plugin 3.5.1 [0].

Fixed an issue in a sample script used to demonstrate MongoDB script execution as the new JavaScript engine seems to be more strict.

Switched to the MongoDB starter and excluded the legacy MongoDB Java driver from projects using Querydsl to consistently make use of the current driver only.

[0] https://issues.apache.org/jira/browse/MCOMPILER-271