#291 - Upgraded to Spring Boot 1.5.6.
Upgraded to Boot 1.5.6. Updated dependencies of reactive projects to latest versions to make sure the APIs of Boot work with the changes in Spring Data Kay RC1. Upgraded all other dependencies to match and got both the Redis and the Cassandra examples to compile again.

Disabled Cassandra reactive samples for now as they fail at runtime.